move meta/index_map/chunk key getters -> hub.util.keys
